+++ src/components/HeaderOrd.tsx
@@ -2,10 +2,7 @@
 import Link from "next/link";
 import Image from "next/image";
 
-import { signIn, signOut, useSession } from "next-auth/react";
-
 import { Menu } from "../stories/Menu/Menu";
-import { Button } from "../stories/Button/Button";
 
 import pathList from "../json/menuPathList.json";
 import menuList from "../json/menuInfoList.json";
@@ -13,13 +10,6 @@ import menuList from "../json/menuInfoList.json";
 import Logo from "../assets/images/logo.png";
 
 export default function Header() {
-	const session = useSession();
-	const signInAction = () => {
-		signIn();
-	}
-	const signOutAction = () => {
-		signOut();
-	}
 	return (
 		<div className="flex items-center justify-between shadow-lg p-6 pb-4 ">
 			<div className="left-box flex items-center">
@@ -45,12 +35,8 @@ export default function Header() {
 					))}
 				</div>
 			</div>
-			<div className="flex">
-				{ session.status === "authenticated" && 
-					<Button label="ログアウト" size="small" className="mr-2 h-[32px]" onClick={() => { signOutAction(); }} /> }
-				{ session.status === "authenticated" ? 
-				<Menu pathList={menuList} size="large" /> :
-				<Button label="ログイン" onClick={() => { signInAction(); }} /> }
+			<div>
+				<Menu pathList={menuList} size="large" />
 			 </div>
 		</div>
 	);
